Move exception catch to handler

+++ WebStoreAPI/Commands/Products/DeleteProductHandler.cs
@@ -1,12 +1,13 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-using MediatR;
+﻿using MediatR;
 using Microsoft.EntityFrameworkCore;
+using System;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace WebStoreAPI.Commands.Products
 {
     //Delete request handler for product
-    public class DeleteProductHandler : IRequestHandler<DeleteProductCommand>
+    public class DeleteProductHandler : IRequestHandler<DeleteProductCommand, DeleteProductCommand>
     {
         private readonly WebStoreContext _context;
 
@@ -15,12 +16,23 @@ namespace WebStoreAPI.Commands.Products
             _context = context;
         }
 
-        public async Task<Unit> Handle(DeleteProductCommand command, CancellationToken cancellationToken)
+        public async Task<DeleteProductCommand> Handle(DeleteProductCommand command, CancellationToken cancellationToken)
         {
-            var product = await _context.Products.FirstOrDefaultAsync(x => x.Id == command.Id, cancellationToken);
-            _context.Products.Remove(product);
-            await _context.SaveChangesAsync(cancellationToken);
-            return Unit.Value;
+            try
+            {
+                var product = await _context.Products.FirstOrDefaultAsync(x => x.Id == command.Id, cancellationToken);
+
+                if (product == null) return null;
+
+                _context.Products.Remove(product);
+                await _context.SaveChangesAsync(cancellationToken);
+                return command;
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+                throw;
+            }            
         }
     }
 }
