+++ WebStoreAPI/Commands/Users/DeleteUserHandler.cs
@@ -1,12 +1,13 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-using MediatR;
+﻿using MediatR;
 using Microsoft.EntityFrameworkCore;
+using System;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace WebStoreAPI.Commands.Users
 {
     //Delete request handler for user
-    public class DeleteUserHandler : IRequestHandler<DeleteUserCommand>
+    public class DeleteUserHandler : IRequestHandler<DeleteUserCommand, DeleteUserCommand>
     {
         private readonly WebStoreContext _context;
 
@@ -15,12 +16,23 @@ namespace WebStoreAPI.Commands.Users
             _context = context;
         }
 
-        public async Task<Unit> Handle(DeleteUserCommand command, CancellationToken cancellationToken)
+        public async Task<DeleteUserCommand> Handle(DeleteUserCommand command, CancellationToken cancellationToken)
         {
-            var user = await _context.Users.FirstOrDefaultAsync(x => x.Id == command.Id, cancellationToken);
-            _context.Users.Remove(user);
-            await _context.SaveChangesAsync(cancellationToken);
-            return Unit.Value;
+            try
+            {
+                var user = await _context.Users.FirstOrDefaultAsync(x => x.Id == command.Id, cancellationToken);
+
+                if (user == null) return null;
+
+                _context.Users.Remove(user);
+                await _context.SaveChangesAsync(cancellationToken);
+                return command;
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine(e);
+                throw;
+            }
         }
     }
 }
